requirements.txt
asgiref==3.3.4
click==8.0.1
Django==3.2.4
Flask==2.0.1
image==1.5.33
itsdangerous==2.0.1
Jinja2==3.0.1
jsonify==0.5
MarkupSafe==2.0.1
numpy==1.20.3
pandas==1.2.4
Pillow==8.2.0
pytesseract==0.3.7
python-dateutil==2.8.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
Werkzeug==2.0.1